first version of main script

+++ SourceCode/SRI/Main/GetAndAnalyseAuditPolicies.psm1
@@ -1,88 +1,26 @@
-Function GetCAPI2 {
-    return wevtutil gl Microsoft-Windows-CAPI2/Operational /f:xml
-}
+Function GetAuditPolicies($importPath) {
+    Write-Host "Get RSoP"
+    $isCurrentPath = $true
+    $pathRSOPXML = $PSScriptRoot + "\rsop.xml"
 
-Function IsCAPI2Enabled([xml] $capi2, [int] $requiredLogSize) {
-    $capi2Enabled = $capi2.channel.enabled
-    $currentLogSize = $capi2.channel.logging.maxsize -as [int]
-    $result = @{}
-
-    if ($capi2Enabled -eq "true" -and $currentLogSize -ge $requiredLogSize) {
-        $result.Add("CAPI2", "EnabledGoodLogSize")
-        $result.Add("CAPI2LogSize", "$currentLogSize")
-        return $result
-    } elseif ($capi2Enabled -eq "true" -and $currentLogSize -lt $requiredLogSize) {
-        $result.Add("CAPI2", "EnabledBadLogSize")
-        $result.Add("CAPI2LogSize", "$currentLogSize")
-        return $result
+    if ($importPath) {
+        $isCurrentPath = $false
+        $pathRSOPXML = $importPath + "\rsop.xml"
     } else {
-        $result.Add("CAPI2", "Disabled")
-        return $result
+        Get-GPResultantSetOfPolicy -ReportType Xml -Path  $pathRSOPXML | Out-Null
     }
-}
 
-Function GetRegistryValue($path, $name) 
-{
-    try {
-        return Get-ItemProperty -Path $path -Name $name -ErrorAction Stop
-    } catch {
-        return $null
-    }
-}
-
-Function IsForceAuditPoliySubcategoryEnabeled($auditPoliySubcategoryKey) {
-    $result = @{}
-
-    if ($auditPoliySubcategoryKey) {
-        if ($auditPoliySubcategoryKey.SCENoApplyLegacyAuditPolicy -eq 1) {
-            $result.Add("ForceAuditPolicySubcategory", "Enabled")
-            return $result
-        } else {
-            $result.Add("ForceAuditPolicySubcategory", "Disabled")
-            return $result
-        }
-    } else {
-        $result.Add("ForceAuditPolicySubcategory", "NotDefined")
-        return $result
-    }
-}
-
-Function GetService($name) {
-    try {
-        return Get-Service -Name $name
-    } catch {
-        throw
-    }
-}
-
-Function IsSysmonInstalled($service) {
-    $result = @{}
-
-    try {
-        if ($service.Status -ne "Running") {
-            $result.Add("Sysmon", "InstalledNotRunning")
-            return $result
-        } else {
-            $result.Add("Sysmon", "InstalledAndRunning")
-            return $result
-        }
-    } catch {
-        $result.Add("Sysmon", "NotInstalled")
-        return $result
-    }
-}
-
-Function GetAuditPolicies {
-    $pathRSOPXML = $PSScriptRoot + "\LocalUserAndComputerReport.xml"
-    
-    Get-GPResultantSetOfPolicy -ReportType Xml -Path  $pathRSOPXML | Out-Null
     $rsopResult = Get-Content $pathRSOPXML
-    Remove-Item $pathRSOPXML
     
+    if ($isCurrentPath) {
+        Remove-Item $pathRSOPXML
+    }
+        
     return $rsopResult
 }
 
 Function AnalyseAuditPolicies ([xml] $rsopResult){    
+    Write-Host "Check RSoP"
     $auditSettingSubcategoryNames = @("Audit Sensitive Privilege Use","Audit Kerberos Service Ticket Operations","Audit Registry","Audit Security Group Management","Audit File System","Audit Process Termination","Audit Logoff","Audit Process Creation","Audit Filtering Platform Connection","Audit File Share","Audit Kernel Object","Audit MPSSVC Rule-Level Policy Change","Audit Non Sensitive Privilege Use","Audit Logon","Audit SAM","Audit Handle Manipulation","Audit Special Logon","Audit Detailed File Share","Audit Kerberos Authentication Service","Audit User Account Management")
     enum AuditSettingValues {
         NoAuditing
@@ -99,6 +37,7 @@ Function AnalyseAuditPolicies ([xml] $rsopResult){
     foreach($auditSettingSubcategoryName in $auditSettingSubcategoryNames) {
         if($auditSettings.SubcategoryName -notcontains $auditSettingSubcategoryName){
             $result.Add(($auditSettingSubcategoryName -replace (" ")), "NotConfigured")
+            Write-Host " - $auditSettingSubcategoryName is not configured" -ForegroundColor Red
         }
     }
 
@@ -137,6 +76,96 @@ Function AnalyseAuditPolicies ([xml] $rsopResult){
     return $result
 }
 
+Function GetRegistryValue($path, $name) 
+{
+    try {
+        return Get-ItemProperty -Path $path -Name $name -ErrorAction Stop
+    } catch {
+        return $null
+    }
+}
+
+Function IsForceAuditPoliySubcategoryEnabeled($auditPoliySubcategoryKey) {
+    Write-Host "Check `'Audit: Force audit policy subcategory settings (Windows Vista or later) to override audit policy category settings`'"
+    $result = @{}
+
+    if ($auditPoliySubcategoryKey) {
+        if ($auditPoliySubcategoryKey.SCENoApplyLegacyAuditPolicy -eq 1) {
+            $result.Add("ForceAuditPolicySubcategory", "Enabled")
+            Write-Host " - `'Audit: Force audit policy subcategory settings (Windows Vista or later) to override audit policy category settings`' enabled" -ForegroundColor Green
+            return $result
+        } else {
+            $result.Add("ForceAuditPolicySubcategory", "Disabled")
+            Write-Host " - `'Audit: Force audit policy subcategory settings (Windows Vista or later) to override audit policy category settings`' disabled" -ForegroundColor Red
+            return $result
+        }
+    } else {
+        $result.Add("ForceAuditPolicySubcategory", "NotDefined")
+        Write-Host " - `'Audit: Force audit policy subcategory settings (Windows Vista or later) to override audit policy category settings`' not defined" -ForegroundColor Red
+        return $result
+    }
+}
+
+Function GetService($name) {
+    try {
+        return Get-Service -Name $name
+    } catch {
+        throw
+    }
+}
+
+Function IsSysmonInstalled($service) {
+    Write-Host "Check Sysmon"
+    $result = @{}
+
+    try {
+        if ($service.Status -ne "Running") {
+            $result.Add("Sysmon", "InstalledNotRunning")
+            Write-Host " - Sysmon is installed but not running as a service" -ForegroundColor Yellow
+            return $result
+        } else {
+            $result.Add("Sysmon", "InstalledAndRunning")
+            Write-Host " - Sysmon is installed and running as a service" -ForegroundColor Green
+            return $result
+        }
+    } catch {
+        $result.Add("Sysmon", "NotInstalled")
+        Write-Host " - Sysmon is not installed" -ForegroundColor Red
+        return $result
+    }
+}
+
+Function GetCAPI2 {
+    return wevtutil gl Microsoft-Windows-CAPI2/Operational /f:xml
+}
+
+Function IsCAPI2Enabled([xml] $capi2, [int] $requiredLogSize) {
+    Write-Host "Check CAPI2"
+    $capi2Enabled = $capi2.channel.enabled
+    $currentLogSize = $capi2.channel.logging.maxsize -as [int]
+    $result = @{}
+    if ($requiredLogSize -lt 4194304) {
+        Write-Host " - Defined Log Size smaller than 4MB ($requiredLogSize) => set default value 4MB (4194304)" -ForegroundColor Yellow
+        $requiredLogSize = 4194304
+    }
+
+    if ($capi2Enabled -eq "true" -and $currentLogSize -ge $requiredLogSize) {
+        $result.Add("CAPI2", "EnabledGoodLogSize")
+        $result.Add("CAPI2LogSize", "$currentLogSize")
+        Write-Host " - CAPI2 enabled with a good log size of $currentLogSize (>= $requiredLogSize)" -ForegroundColor Green
+        return $result
+    } elseif ($capi2Enabled -eq "true" -and $currentLogSize -lt $requiredLogSize) {
+        $result.Add("CAPI2", "EnabledBadLogSize")
+        $result.Add("CAPI2LogSize", "$currentLogSize")
+        Write-Host " - CAPI2 enabled with a bad log size of $currentLogSize (<= $requiredLogSize)" -ForegroundColor Red
+        return $result
+    } else {
+        $result.Add("CAPI2", "Disabled")
+        Write-Host " - CAPI2 disabled" -ForegroundColor Red
+        return $result
+    }
+}
+
 Function MergeHashtables {
     $Output =  [ordered]@{}
     ForEach ($Hashtable in ($Input + $Args)) {
@@ -155,9 +184,14 @@ Function WriteXMLElement([System.XMl.XmlTextWriter] $XmlWriter, [String] $startE
     $xmlWriter.WriteEndElement()
 }
 
-Function WriteXML($resultCollection) {
+Function WriteXML($resultCollection, $exportPath) {
+    Write-Host "Write Result XML"
     $resultXML = $PSScriptRoot + "\resultOfAuditPolicies.xml"
+    if ($exportPath) {
+        $resultXML = $exportPath + "\resultOfAuditPolicies.xml"
+    }
     $xmlWriter = New-Object System.XMl.XmlTextWriter($resultXML, $Null)
+
     $xmlWriter.Formatting = "Indented"
     $xmlWriter.Indentation = 1
     $xmlWriter.IndentChar = "`t"
@@ -172,6 +206,7 @@ Function WriteXML($resultCollection) {
     $xmlWriter.WriteEndDocument()
     $xmlWriter.Flush()
     $xmlWriter.Close()
+    Write-Host "DONE!!!"
 }
 
-Export-ModuleMember -Function GetCAPI2, IsCAPI2Enabled, GetRegistryValue, IsForceAuditPoliySubcategoryEnabeled, GetService, IsSysmonInstalled, GetAuditPolicies, AnalyseAuditPolicies, MergeHashtables, WriteXML
+Export-ModuleMember -Function GetAuditPolicies, AnalyseAuditPolicies, GetRegistryValue, IsForceAuditPoliySubcategoryEnabeled, GetService, IsSysmonInstalled, GetCAPI2, IsCAPI2Enabled, MergeHashtables, WriteXML
